requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-weaviate
langserve
sentence-transformers
weaviate-client
python-dotenv
fastapi
uvicorn
pydantic
torch
transformers
huggingface-hub
langchain-core
python-multipart


